chore: update test configuration for storybook
Update the test configuration in the storybook.ts file to use the 'next' branch of the 'storybookjs/sandboxes' repository. This change ensures that the correct test subdirectory is used for the React Vite default TypeScript setup. Additionally, the 'beforeTest' and 'test' functions have been updated to use the 'getCommand' function from the '@antfu/ni' package for executing commands. These changes improve the reliability and flexibility of the test setup and considers the used package manager